requirements.txt
asttokens==2.2.1
async-timeout==4.0.2
backcall==0.2.0
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
comm==0.1.2
cryptography==41.0.2
debugpy==1.6.5
decorator==5.1.1
distlib==0.3.6
entrypoints==0.4
executing==1.2.0
filelock==3.12.2
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.20.2
ipython==8.8.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==7.4.9
jupyter_core==5.1.3
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mysql==0.0.3
mysqlclient==2.2.0
nest-asyncio==1.5.6
numpy==1.26.1
packaging==23.0
paramiko==3.3.1
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.8.1
prompt-toolkit==3.0.36
psutil==5.9.4
pure-eval==0.2.2
pycparser==2.21
Pygments==2.14.0
PyMySQL==1.1.0
PyNaCl==1.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pywin32==305
pyzmq==25.0.0
redis==4.6.0
requests==2.31.0
six==1.16.0
spotipy==2.23.0
SQLAlchemy==2.0.19
sshtunnel==0.4.0
stack-data==0.6.2
tornado==6.2
traitlets==5.8.1
typing_extensions==4.7.1
urllib3==2.0.4
virtualenv==20.23.1
wcwidth==0.2.6
Werkzeug==2.3.6
zipp==3.16.2
